refactoring to signature generation logic
